Move HttpBackendProvider to httpUtils
